+++ lineapy/cli/cli.py
@@ -25,8 +25,8 @@ from nbconvert.preprocessors import ExecutePreprocessor
 from rich.console import Console
 from rich.progress import Progress
 
-from lineapy.api.api_classes import LineaArtifact
 from lineapy.api.api_utils import extract_taskgraph
+from lineapy.api.models.linea_artifact import LineaArtifact
 from lineapy.data.types import SessionType
 from lineapy.db.db import RelationalLineaDB
 from lineapy.exceptions.excepthook import set_custom_excepthook
